Split config sections from panel config, add CI for entrypoint size
